refactor: move deepIsEqual to /common
so it can be used on backend as well
