+++ src/common/utils.ts
@@ -1,3 +1,5 @@
+import isEqual from "lodash/isEqual";
+
 export const getBaseUrl = () => {
   if (typeof window !== "undefined")
     // browser should use relative path
@@ -6,4 +8,8 @@ export const getBaseUrl = () => {
   if (!process.env.BASE_URL) throw new Error("Missing BASE_URL env var");
 
   return process.env.BASE_URL;
+};
+
+export const deepIsEqual = (a: unknown, b: unknown) => {
+  return isEqual(a, b);
 };
