+++ src/web/view/components/Filter/TableFilters.tsx
@@ -3,9 +3,9 @@ import { Stack } from "@mui/material";
 import { useCallback, useEffect } from "react";
 import { FormProvider, useForm } from "react-hook-form";
 
+import { deepIsEqual } from "../../../../common/utils";
 import { FormContext } from "../../../common/components/Form/FormContext";
 import { NodeSelect } from "../../../common/components/Form/NodeSelect";
-import { deepIsEqual } from "../../../common/store/utils";
 import { useCriteria, useNodesOfType, useSolutions } from "../../../topic/store/nodeHooks";
 import { setTableFilter, useTableFilterWithFallbacks } from "../../currentViewStore/filter";
 import { TableFilter, tableFilterSchema } from "../../utils/tableFilter";
